added Debug mode to the Module
remove the environment variable for the debug file, added some utilities to use callback or promises inside of the controller, I haven't test this I will do it whenever the testing branch is merged.

we don't use notification attributes because we don't know what attributes the notification model will have.